Entrega_Final/requirements.txt
pandas
requests
sqlalchemy
apache-airflow